% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{delta}
\alias{delta}
\title{delta}
\usage{
delta(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
The function delta(\code{a},\code{b}), if \code{a} == \code{b}, return 1; else, return 0.
}
\description{
the delta function
}

