% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeCell.R
\name{InitializeCell}
\alias{InitializeCell}
\title{InitializeCell}
\usage{
InitializeCell(y, param, initial, const)
}
\arguments{
\item{y}{A one-row matrix, the new sample to be input.}

\item{param}{A logical scalar, if TRUE, the model is in parametric version, otherwise, a non-parametric one.}

\item{initial}{A numeric vector, specifies the initial value of parameters over the continuous domain, if param = T, initial = c(pi_1,mean_1,cov_1, ..., pi_K, mean_K,cov_K), if param = F, initial = c(q1,q2, ..., qK).}

\item{const}{A numeric vector, specifies the value of all global variables, if param = T, then const = c(N,n,d,rh,r,K,alpha); if param=FALSE, then const = c(N,n,d,rh,r,K,sigma_sqare).}
}
\value{
The matrix which stores updated parameters over the continuous domain.
}
\description{
initializes parameters in the continuous domain while inputting the first sample
}

