% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputSample.R
\name{InputSample}
\alias{InputSample}
\title{InputSample}
\usage{
InputSample(y, param = TRUE, smart, const, initial)
}
\arguments{
\item{y}{A matrix, the new sample to be input.}

\item{param}{A logical scalar, if TRUE, the model is in parametric version, otherwise, a non-parametric one.}

\item{smart}{A matrix, stores all the parameters over the continuous domain.}

\item{const}{A numeric vector, specifies the value of all global variables, if param = T, then const = c(N,n,d,rh,r,K,alpha); if param=FALSE, then const = c(N,n,d,rh,r,K,sigma_sqare).}

\item{initial}{A numeric vector, specifies the initial value of parameters over the continuous domain, if param = T, initial = c(pi_1,mean_1,cov_1, ..., pi_K, mean_K,cov_K), if param = F, initial = c(q1,q2, ..., qK).}
}
\value{
The matrix which stores updated parameters over the continuous domain.
}
\description{
updates parameters after inputting sample (can be more than one)
}

