% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateConst.R
\name{UpdateConst}
\alias{UpdateConst}
\title{UpdateConst}
\usage{
UpdateConst(y, const)
}
\arguments{
\item{y}{A matrix, the new sample to be input.}

\item{const}{A numeric vector, specifies the value of all global variables, if param = T, then const = c(N,n,d,rh,r,K,alpha); if param=FALSE, then const = c(N,n,d,rh,r,K,sigma_sqare).}
}
\value{
The updated vector which specifies all the constant parameters.
}
\description{
updates the vector const after inputting sample y
}

