% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhichCell.R
\name{WhichCell}
\alias{WhichCell}
\title{WhichCell}
\usage{
WhichCell(y, param, smart, const)
}
\arguments{
\item{y}{A one-row matrix, the new sample to be input.}

\item{param}{A logical scalar, if TRUE, the model is in parametric version, otherwise, a non-parametric one.}

\item{smart}{A matrix, stores all the parameters over the continuous domain.}

\item{const}{A numeric vector, specifies the value of all global variables, if param = T, then const = c(N,n,d,rh,r,K,alpha); if param=FALSE, then const = c(N,n,d,rh,r,K,sigma_sqare).}
}
\value{
The row index of the discrete class to which the new sample belongs.
}
\description{
returns the index of the cell to which the new sample belongs
}

